#95 Removed style.css
