+++ src/components/molecules/BookItem/index.tsx
@@ -4,23 +4,42 @@ import H4 from "../../atoms/H4";
 import FastImage from "react-native-fast-image";
 import { Book } from "@sqd-ffrog/services";
 import H6 from "../../atoms/H6";
+import { LightColor } from "../../atoms/colors";
 
 interface BookItemProps {
   item: Book;
 }
 
 function BookItem({
-  item: { title, thumbnailUri, author, isbn, publisher }
+  item: { title, thumbnailUri, author, isbn, publisher, branchVolumes }
 }: BookItemProps) {
   return (
     <TouchableOpacity key={isbn} style={styles.containerStyle}>
       <FastImage source={{ uri: thumbnailUri }} style={styles.imageStyle} />
       <View style={styles.innerContainerStyle}>
         <H4 style={styles.titleStyle} numberOfLines={1}>
-          {title + "sfsfsfdfasdfafasfsfsafsaf"}
+          {title}
         </H4>
         <H6>{author}</H6>
         <H6>{publisher}</H6>
+        {branchVolumes.length > 0 && (
+          <View style={styles.branchVolumeStyle}>
+            <H6 style={styles.volumeTextStyle}>
+              {`${branchVolumes[0].location}[${branchVolumes[0].volume}]`}
+            </H6>
+            <H6
+              // TODO: 여기 color 나중에 다른로직 쓰기. hasItem, isSubscribed 혼합해서
+              style={{
+                color: /가능/.exec(branchVolumes[0].state)
+                  ? LightColor.accentColor
+                  : LightColor.mainColor,
+                ...styles.volumeStateTextStyle
+              }}
+            >
+              {branchVolumes[0].state}
+            </H6>
+          </View>
+        )}
       </View>
     </TouchableOpacity>
   );
@@ -38,11 +57,22 @@ const styles = StyleSheet.create({
     flex: 1
   },
   imageStyle: {
-    width: 60,
-    height: 80
+    width: 75,
+    height: 100
   },
   titleStyle: {
     fontWeight: "700"
+  },
+  branchVolumeStyle: {
+    flexDirection: "row",
+    justifyContent: "space-between"
+  },
+  volumeTextStyle: {
+    flex: 1
+  },
+  volumeStateTextStyle: {
+    fontWeight: "700",
+    marginHorizontal: 8
   }
 });
