+++ src/components/molecules/DreamyTile/index.tsx
@@ -1,7 +1,7 @@
 import React from "react";
+import { StyleSheet } from "react-native";
 import Icon from "react-native-vector-icons/AntDesign";
 import { H6, LightColor } from "@sqd-ffrog/components";
-import styles from "./styles";
 import { TouchableOpacity } from "react-native";
 import { normalize } from "@sqd-ffrog/tools";
 
@@ -23,5 +23,20 @@ function DreamyTile({ name, title }: DreamyTileProps) {
     </TouchableOpacity>
   );
 }
+
+const styles = StyleSheet.create({
+  containerStyle: {
+    flexDirection: "row",
+    alignItems: "center",
+    marginHorizontal: 16,
+    marginVertical: 12
+  },
+  iconStyle: {
+    marginRight: 12
+  },
+  titleStyle: {
+    fontWeight: "700"
+  }
+});
 
 export default DreamyTile;
