--- src/components/molecules/DreamyTile/styles.tsx
@@ -1,18 +1,0 @@
-import { StyleSheet } from "react-native";
-
-const styles = StyleSheet.create({
-  containerStyle: {
-    flexDirection: "row",
-    alignItems: "center",
-    marginHorizontal: 16,
-    marginVertical: 12
-  },
-  iconStyle: {
-    marginRight: 12
-  },
-  titleStyle: {
-    fontWeight: "700"
-  }
-});
-
-export default styles;
