+++ src/components/molecules/LibrarySeatItem/index.tsx
@@ -1,8 +1,9 @@
 import React from "react";
+import { StyleSheet } from "react-native";
 import { LibrarySeatProps } from "@sqd-ffrog/react-native-jnu/src/library";
 import { View } from "react-native";
 import H6 from "../../atoms/H6";
-import styles from "./styles";
+import { LightColor } from "../../atoms/colors";
 
 function LibrarySeatItem({
   id,
@@ -18,5 +19,20 @@ function LibrarySeatItem({
     </View>
   );
 }
+
+const styles = StyleSheet.create({
+  containerStyle: {
+    marginHorizontal: 16,
+    marginVertical: 8,
+    flexDirection: "row",
+    justifyContent: "space-between",
+    alignItems: "center"
+  },
+  seatTextStyle: {
+    fontWeight: "700",
+    color: LightColor.mainColor,
+    textAlign: "center"
+  }
+});
 
 export default LibrarySeatItem;
