+++ src/components/molecules/MainRightHeaderTitle/index.tsx
@@ -1,8 +1,7 @@
 import React from "react";
-import { View } from "react-native";
+import { View, StyleSheet } from "react-native";
 import { SettingsButton, UserButton, LightColor } from "@sqd-ffrog/components";
 import { useNavigation } from "@react-navigation/native";
-import styles from "./styles";
 
 function MainRightHeaderTitle() {
   const navigation = useNavigation();
@@ -15,5 +14,12 @@ function MainRightHeaderTitle() {
     </View>
   );
 }
+
+const styles = StyleSheet.create({
+  style: {
+    flexDirection: "row",
+    marginHorizontal: 8
+  }
+});
 
 export default MainRightHeaderTitle;
