--- src/components/molecules/MainRightHeaderTitle/styles.tsx
@@ -1,10 +1,0 @@
-import { StyleSheet } from "react-native";
-
-const styles = StyleSheet.create({
-  style: {
-    flexDirection: "row",
-    marginHorizontal: 8
-  }
-});
-
-export default styles;
