--- src/components/molecules/NewBookEmpty/styles.tsx
@@ -1,10 +1,0 @@
-import { StyleSheet } from "react-native";
-
-const styles = StyleSheet.create({
-  textStyle: {
-    textAlign: "center",
-    marginVertical: 16
-  }
-});
-
-export default styles;
