--- src/components/molecules/NewBooksItem/styles.tsx
@@ -1,19 +1,0 @@
-import { StyleSheet } from "react-native";
-import { LightColor } from "../../atoms/colors";
-
-const styles = StyleSheet.create({
-  containerStyle: {
-    maxWidth: 90,
-    marginHorizontal: 8
-  },
-  imageStyle: {
-    width: 90,
-    height: 120
-  },
-  titleStyle: {
-    fontWeight: "700",
-    marginVertical: 4
-  }
-});
-
-export default styles;
