--- src/components/molecules/SearchBar/styles.tsx
@@ -1,17 +1,0 @@
-import { StyleSheet } from "react-native";
-
-const styles = StyleSheet.create({
-  cardStyle: {
-    flexDirection: "row",
-    alignItems: "center",
-    marginBottom: 16
-  },
-  inputStyle: {
-    flex: 1
-  },
-  closeStyle: {
-    marginHorizontal: 4
-  }
-});
-
-export default styles;
