--- src/components/molecules/SearchTopTab/styles.tsx
@@ -1,15 +1,0 @@
-import { StyleSheet } from "react-native";
-
-const styles = StyleSheet.create({
-  contentContainerStyle: {
-    marginHorizontal: 16,
-    marginVertical: 8,
-  },
-  tabTextStyle: {
-    fontWeight: "700",
-    marginHorizontal: 8,
-    marginVertical: 4
-  }
-});
-
-export default styles;
