+++ src/components/molecules/Title/index.tsx
@@ -1,6 +1,6 @@
 import React from "react";
+import { StyleSheet } from "react-native";
 import { H1 } from "@sqd-ffrog/components";
-import styles from "./styles";
 
 interface TitleProps {
   children: string;
@@ -9,5 +9,13 @@ interface TitleProps {
 function Title({ children }: TitleProps) {
   return <H1 style={styles.titleStyle}>{children}</H1>;
 }
+
+const styles = StyleSheet.create({
+  titleStyle: {
+    marginHorizontal: 16,
+    marginVertical: 8,
+    fontWeight: "700"
+  }
+});
 
 export default Title;
