--- src/components/molecules/Title/styles.tsx
@@ -1,11 +1,0 @@
-import { StyleSheet } from "react-native";
-
-const styles = StyleSheet.create({
-  titleStyle: {
-    marginHorizontal: 16,
-    marginVertical: 8,
-    fontWeight: "700"
-  }
-});
-
-export default styles;
