--- src/components/organisms/DreamyTotalMenu/styles.tsx
@@ -1,12 +1,0 @@
-import { StyleSheet } from "react-native";
-
-const styles = StyleSheet.create({
-  titleStyle: {
-    marginTop: 24
-  },
-  tileContainerStyle: {
-    flexDirection: "column"
-  }
-});
-
-export default styles;
