--- src/components/organisms/FoodCardCarousel/styles.tsx
@@ -1,9 +1,0 @@
-import { StyleSheet } from "react-native";
-
-const styles = StyleSheet.create({
-  wrapper: {
-    height: 240
-  }
-});
-
-export default styles;
