--- src/components/organisms/LibraryNewBook/styles.tsx
@@ -1,10 +1,0 @@
-import { StyleSheet } from "react-native";
-
-const styles = StyleSheet.create({
-  contentContainerStyle: {
-    marginHorizontal: 16,
-    marginBottom: 24
-  }
-});
-
-export default styles;
