+++ src/view/LibraryBook/index.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+import { StyleSheet } from "react-native";
+import { ScrollView } from "@sqd-ffrog/components";
+
+function LibraryBook() {
+  return <ScrollView title="도서상세"></ScrollView>;
+}
+
+const styles = StyleSheet.create({});
+
+export default LibraryBook;
